fix: security hardening, RBAC, pagination, infinite scroll, and test fixes
Security:
- Add global request body size limit (5MB middleware)
- Add rate limiting (10/min) to auth endpoints
- Add configurable CORS allowed origin (ALLOWED_ORIGIN env)
- Escape LIKE wildcards (% and _) in employee search query
- Enforce minimum 16-char JWT secret, lower default token TTL to 72h
- Add RBAC checks to all 16 write endpoints (block employee role)
- Validate wage_type enum (daily/monthly/hourly) on staff create
- Propagate strconv.ParseFloat errors instead of silent zero defaults
- Add date format validation (YYYY-MM-DD) across all controllers
- Add max length limits on name/phone fields
- Fix empty createdBy in shift activity logs (pass claims.EmployeeID)
- Rename grace_minutes JSON field to grace_period_minutes

Frontend pagination & infinite scroll:
- Convert shifts_management_page to ConsumerStatefulWidget with infinite scroll
- Convert holidays_page to ConsumerStatefulWidget with infinite scroll
- Add ledgerRefreshProvider for cross-page invalidation
- Pass limit params to attendance/holiday/shift providers

Code quality:
- Standardize JSONFail for 4xx / JSONError for 5xx across all 12 controllers
- Fix all pre-existing broken mock expectations in tests
  (missing CreateActivityLog, FindEmployeeByID, FindTenantByPhone,
   int32->float64 type mismatches)
- All 18 test suites pass, flutter analyze 0 issues

+++ server/controllers/api/v1/ledger_controller_test.go
@@ -82,7 +82,7 @@ func TestLedgerGetBalance_Success(t *testing.T) {
 
 	mockQuerier.EXPECT().
 		GetBalanceByEmployee(gomock.Any(), gomock.Any()).
-		Return(int32(1500), nil)
+		Return(float64(1500), nil)
 
 	r := chi.NewRouter()
 	r.Get("/api/v1/ledger/{id}/balance", ctrl.GetBalance)
@@ -270,7 +270,7 @@ func TestLedgerGetBalance_DBError(t *testing.T) {
 
 	mockQuerier.EXPECT().
 		GetBalanceByEmployee(gomock.Any(), gomock.Any()).
-		Return(int32(0), errors.New("db error"))
+		Return(float64(0), errors.New("db error"))
 
 	r := chi.NewRouter()
 	r.Get("/api/v1/ledger/{id}/balance", ctrl.GetBalance)
